Tidy output to stdout
