+++ README.md
@@ -87,7 +87,7 @@ Once running, the following keys are mapped to actions:
     ' ' - pause/resume playback
     'c' - display list of channels and current events to the console
         - Also shows basic onscreen channel listing, whilst onscreen use
-        - n for next screen of channels, or p for previous screen of channels in listing
+        - d for down a screen of channels, or u for up a screen of channels in listing
     'o' - To toggle subscribe/unsubscribe (idle) on current channel
     'a' - Cycle through available audio streams
     'z' - Force 4:3 (pillarbox) or 16:9 (fullscreen) display
