+++ channels.c
@@ -6,6 +6,24 @@
 
 /* Channels are merged from different servers, based on LCN */
 
+struct channel_t
+{
+  int id;   /* Our internal ID */
+  int tvh_id[MAX_HTSP_SERVERS]; /* tvheadend ID */
+  uint32_t eventId[MAX_HTSP_SERVERS];
+  uint32_t nextEventId[MAX_HTSP_SERVERS];
+  int lcn;
+  int type;
+  char* name;
+  struct channel_t* next;
+  struct channel_t* prev;
+};
+
+static struct channel_t* channels;
+static struct channel_t* channels_cache;
+
+static int next_id = 0;
+
 void channels_init(void)
 {
   channels = NULL;
@@ -95,11 +113,6 @@ void channels_update(int server, int lcn, int tvh_id, char* name, int type, uint
     if (nextEventId) p->nextEventId[server] = nextEventId;
   }
 }
-
-struct channel_t* channels_return_struct(void)
-{
-return channels;
-};
 
 void channels_dump(void)
 {
